refactor: move TextExtractionStatus and related types to @vertesia/common
Move types that zeno-server needs out of @vertesia/workflow so zeno-server
can drop the heavy workflow dependency (sharp, tiktoken) from its Docker image.

Moved types: TextExtractionResult, TextExtractionStatus, TranscriptMediaResult
Removed re-exports from result-types.ts and barrel files.
Updated all internal imports in @vertesia/workflow to use @vertesia/common.
