+++ packages/workflow/src/dsl.ts
@@ -1,4 +1,3 @@
 export * from "./dsl/dslProxyActivities.js";
 export * from "./errors.js";
-export * from "./result-types.js";
