+++ packages/workflow/src/index.ts
@@ -30,7 +30,6 @@ export * from "./iterative-generation/activities/index.js";
 
 export * from "./dsl/setup/ActivityContext.js";
 export * from "./errors.js";
-export * from "./result-types.js";
 export * from "./utils/blobs.js";
 export * from "./utils/client.js";
 export * from "./utils/memory.js";
